% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn.R
\name{finn}
\alias{finn}
\title{Create table from Finn search}
\usage{
finn(area = 1:5, pmin = 3, pmax = 5, add = NULL, ...)
}
\arguments{
\item{area}{Areas to choose in Oslo}

\item{pmin}{Minimum price}

\item{pmax}{Maximum price}

\item{add}{Other areas code which isn't included in the default}
}
\description{
Create table from Finn search
}
