% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_more.R
\name{get_add}
\alias{get_add}
\title{Adresse or postcode}
\usage{
get_add(x, postcode = FALSE)
}
\arguments{
\item{x}{HTML code}

\item{postcode}{Get only postcode. Default is the whole address}
}
\description{
Adresse or postcode
}
