% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_database.R
\name{make_db}
\alias{make_db}
\title{Create SQLite database}
\usage{
make_db(x, ...)
}
\arguments{
\item{x}{Data table from \link{finn} function}

\item{...}{Extra arguments}
}
\description{
This function will create a \strong{boligfinn} folder in the HOME directory
if it isn't there and create a \emph{boligfinn.sqlite} database.
}
