% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn.R
\name{finn}
\alias{finn}
\title{Create table from Finn search}
\usage{
finn(area = 1, pmin = 35, pmax = 5, add = NULL, ...)
}
\arguments{
\item{area}{Areas in Oslo to choose from
\itemize{
  \item 1:Bygdøy/Frogner 
  \item 2:Grünerløkka/Sofienberg 
  \item 3:Sagene/Torshov
  \item 4:St.Hanshaugen/Ullevål 
  \item 5:Uranienborg/Majorstuen
 }}

\item{pmin}{Minimum price}

\item{pmax}{Maximum price}

\item{add}{Other areas to choose from which isn't included in the default}

\item{...}{Extra arguments}
}
\description{
Create table from Finn search
}
