% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_finn.R
\name{get_finn}
\alias{get_finn}
\title{Get all the Finn kodes}
\usage{
get_finn(html = NULL)
}
\arguments{
\item{html}{\code{HTML} search from
\url[Finn.no]{https://www.finn.no/realestate/browse.html}}
}
\description{
Get all the Finn kodes
}
