% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_finn.R
\name{get_flat}
\alias{get_flat}
\title{Get detail for a coded appartment}
\usage{
get_flat(code)
}
\arguments{
\item{code}{The code for a specific appartment or house}
}
\description{
Gether all the detail information from the selected flat such as flat type,
size etc.
}
