% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html.R
\name{get_html}
\alias{get_html}
\title{Get HTML codes}
\usage{
get_html(area, pmin, pmax, add)
}
\arguments{
\item{area}{Areas in Oslo to choose from
1:Bygdøy/Frogner 2:Grünerløkka/Sofienberg 3:Sagene/Torshov
4:St.Hanshaugen/Ullevål 5:Uranienborg/Majorstuen}

\item{pmin}{Minimum price}

\item{pmax}{Maximum price}

\item{add}{Other areas to choose from which isn't included in the default}
}
\description{
Collect all \code{HTML} code for further anaysis
}
