% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html.R
\name{get_html}
\alias{get_html}
\title{Get HTML codes}
\usage{
get_html(area = 1:5, pmin = 3, pmax = 5, add = NULL, ...)
}
\arguments{
\item{area}{Areas to choose in Oslo}

\item{pmin}{Minimum price}

\item{pmax}{Maximum price}

\item{add}{Other areas code which isn't included in the default}
}
\description{
Collect all \code{HTML} code for further anaysis
}
